+++ api/src/main/java/com/codinglitch/lexiconfig/LexiconEvent.java
@@ -1,0 +1,4 @@
+package com.codinglitch.lexiconfig;
+
+public class LexiconEvent {
+}