+++ api/src/main/java/com/codinglitch/lexiconfig/LexiconfigApi.java
@@ -9,10 +9,13 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Consumer;
 
 public abstract class LexiconfigApi {
-    protected static final List<LexiconData> PRE_REGISTERED_LEXICONS = new ArrayList<>();
-    protected static final Map<Event, Runnable> LISTENERS = new HashMap<>();
+    protected static final List<LexiconData> SHELVED_LEXICONS = new ArrayList<>();
+    protected static final Map<Consumer<LexiconEvent>, EventType> LISTENERS = new HashMap<>();
+
+    public static final List<Library> LIBRARIES = new ArrayList<>();
 
     public static LexiconfigApi INSTANCE;
 
@@ -20,12 +23,10 @@ public abstract class LexiconfigApi {
         INSTANCE = this;
     }
 
-    public enum Event {
-        STARTUP,
-        RELOAD
-    }
-
-    private static Logger LOGGER = LogManager.getLogger(LexiconfigApi.class);
+    /**
+     * Do not use these.
+     */
+    private static Logger LOGGER = LogManager.getLogger("Lexiconfig");
     public static void info(Object object, Object... substitutions) {
         LOGGER.info(String.valueOf(object), substitutions);
     }
@@ -39,19 +40,49 @@ public abstract class LexiconfigApi {
         LOGGER.error(String.valueOf(object), substitutions);
     }
 
-    protected abstract void registerLexicon(LexiconData lexicon);
-    public static void register(LexiconData lexicon) {
-        if (LexiconfigApi.INSTANCE != null) {
-            LexiconfigApi.INSTANCE.registerLexicon(lexicon);
-            return;
-        }
+    /**
+     * These are all the available event types for use inside the {@code registerListener} method.
+     * @see com.codinglitch.lexiconfig.LexiconfigApi#registerListener(EventType, Consumer)
+     */
+    public enum EventType {
+        PRE_CATALOG,
+        POST_CATALOG,
 
-        PRE_REGISTERED_LEXICONS.add(lexicon);
+        PRE_REVISION,
+        POST_REVISION,
+        PRE_LEXICON_REVISION,
+        POST_LEXICON_REVISION
     }
 
-    public static void registerListener(Event eventType, Runnable listener) {
-        LISTENERS.put(eventType, listener);
+    /**
+     * This method is used to shelve a lexicon to be registered for reloading events, etc. and should be called within the {@code shelveLexicons} method of a {@code LexiconLibrary}.
+     * @see Library
+     * @see Library#shelveLexicons
+     * @param lexicon The lexicon to shelve
+     */
+    public static void shelveLexicon(LexiconData lexicon) {
+        SHELVED_LEXICONS.add(lexicon);
+        info("Shelved lexicon {}!", lexicon);
     }
 
+    /**
+     * This used to register a listener of a certain type, which is fired in various parts of the lifecycle.
+     * @see com.codinglitch.lexiconfig.LexiconfigApi.EventType
+     * @see com.codinglitch.lexiconfig.events
+     */
+    public static <E extends LexiconEvent> void registerListener(EventType eventType, Consumer<E> listener) {
+        LISTENERS.put((Consumer<LexiconEvent>) listener, eventType);
+    }
+
+    protected void callEvent(EventType eventType, LexiconEvent event) {
+        LISTENERS.forEach((runnable, type) -> {
+            if (type == eventType) runnable.accept(event);
+        });
+    }
+
+    /**
+     * This is used to retrieve the configuration folder path, dependent on the modloader.
+     * @return The path of the config folder
+     */
     public abstract Path getConfigPath();
 }