+++ common/src/main/java/com/codinglitch/lexiconfig/Lexiconfig.java
@@ -1,8 +1,9 @@
 package com.codinglitch.lexiconfig;
 
 import com.codinglitch.lexiconfig.classes.LexiconData;
+import com.codinglitch.lexiconfig.events.RevisionEvent;
+import com.codinglitch.lexiconfig.events.CatalogEvent;
 import com.codinglitch.lexiconfig.platform.Services;
-import com.codinglitch.lexiconfig.LexiconfigApi;
 
 import java.nio.file.Path;
 import java.util.*;
@@ -10,8 +11,6 @@ import java.util.*;
 public class Lexiconfig extends LexiconfigApi {
     public static final Lexiconfig API = new Lexiconfig();
     public static final String ID = "lexiconfig";
-
-    private static final List<LexiconData> REGISTERED_LEXICONS = new ArrayList<>();
 
     public static <T> T loadService(Class<T> clazz) {
         final T loadedService = ServiceLoader.load(clazz)
@@ -22,30 +21,41 @@ public class Lexiconfig extends LexiconfigApi {
     }
 
     @Override
-    public void registerLexicon(LexiconData lexicon) {
-        REGISTERED_LEXICONS.add(lexicon);
-
-        Lexiconfig.info("Registering lexicon {}", lexicon.getClass().getName());
-
-        lexicon.load();
-        lexicon.save();
-    }
-
-    @Override
     public Path getConfigPath() {
         return Services.PLATFORM.getConfigPath();
     }
 
+    public static void revise() {
+        info("Starting lexicon revision!");
+        API.callEvent(EventType.PRE_REVISION, new RevisionEvent());
 
-    public void callEvent(Event eventType) {
-        LISTENERS.forEach((event, runnable) -> {
-            if (event == eventType) runnable.run();
-        });
+        for (LexiconData lexicon : SHELVED_LEXICONS) {
+            API.callEvent(EventType.PRE_LEXICON_REVISION, new RevisionEvent.Lexicon(lexicon));
+
+            lexicon.load();
+            lexicon.save();
+
+            API.callEvent(EventType.POST_LEXICON_REVISION, new RevisionEvent.Lexicon(lexicon));
+        }
+
+        API.callEvent(EventType.POST_REVISION, new RevisionEvent());
+        info("Finished lexicon revision!");
     }
 
     public static void initialize() {
-        PRE_REGISTERED_LEXICONS.forEach(API::registerLexicon);
-        API.callEvent(Event.STARTUP);
-        API.callEvent(Event.RELOAD);
+        Services.PLATFORM.shelveLexicons();
+    }
+
+    public static void postInitialize() {
+        API.callEvent(EventType.PRE_CATALOG, new CatalogEvent());
+
+        for (LexiconData lexicon : SHELVED_LEXICONS) {
+            info("Cataloging lexicon {}!", lexicon);
+            lexicon.load();
+            lexicon.save();
+        }
+
+        API.callEvent(EventType.POST_CATALOG, new CatalogEvent());
+        info("Lexiconfig cataloging completed!");
     }
 }