+++ fabric/src/main/java/com/codinglitch/lexiconfig/FabricLexiconfig.java
@@ -7,5 +7,7 @@ public class FabricLexiconfig implements ModInitializer {
     @Override
     public void onInitialize() {
         Lexiconfig.initialize();
+        Lexiconfig.postInitialize();
     }
+
 }